+++ lib/Server.js
@@ -4,10 +4,7 @@
   import/order,
   no-shadow,
   no-undefined,
-  func-names,
-  multiline-ternary,
-  array-bracket-spacing,
-  space-before-function-paren
+  func-names
 */
 const fs = require('fs');
 const path = require('path');
@@ -702,7 +699,10 @@ class Server {
       return true;
     }
 
-    if (!headerToCheck) headerToCheck = 'host';
+    if (!headerToCheck) {
+      headerToCheck = 'host';
+    }
+
     // get the Host header and extract hostname
     // we don't care about port not matching
     const hostHeader = headers[headerToCheck];
@@ -736,7 +736,9 @@ class Server {
       for (let hostIdx = 0; hostIdx < this.allowedHosts.length; hostIdx++) {
         const allowedHost = this.allowedHosts[hostIdx];
 
-        if (allowedHost === hostname) return true;
+        if (allowedHost === hostname) {
+          return true;
+        }
 
         // support "." as a subdomain wildcard
         // e.g. ".example.com" will allow "example.com", "www.example.com", "subdomain.example.com", etc
@@ -778,7 +780,7 @@ class Server {
   listen(port, hostname, fn) {
     this.hostname = hostname;
 
-    const returnValue = this.listeningApp.listen(port, hostname, (err) => {
+    return this.listeningApp.listen(port, hostname, (err) => {
       const socket = sockjs.createServer({
         // Use provided up-to-date sockjs-client
         sockjs_url: '/__webpack_dev_server__/sockjs.bundle.js',
@@ -849,8 +851,6 @@ class Server {
         fn.call(this.listeningApp, err);
       }
     });
-
-    return returnValue;
   }
 
   close(cb) {
