+++ lib/utils/addEntries.js
@@ -1,11 +1,5 @@
 'use strict';
 
-/* eslint-disable
-  no-shadow,
-  no-param-reassign,
-  array-bracket-spacing,
-  space-before-function-paren
-*/
 const webpack = require('webpack');
 const createDomain = require('./createDomain');
 
@@ -50,6 +44,7 @@ function addEntries(config, options, server) {
       return entries.concat(entry);
     };
 
+    // eslint-disable-next-line no-shadow
     [].concat(config).forEach((config) => {
       config.entry = prependEntry(config.entry || './src');
