+++ lib/utils/createCertificate.js
@@ -1,8 +1,5 @@
 'use strict';
 
-/* eslint-disable
-  space-before-function-paren
-*/
 const selfsigned = require('selfsigned');
 
 function createCertificate(attrs) {
