+++ lib/utils/createConfig.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const { defaultTo } = require('../../bin/utils');
+const defaultTo = require('./defaultTo');
 
 function createConfig(config, argv, { port }) {
   const firstWpOpt = Array.isArray(config) ? config[0] : config;
