+++ lib/utils/createDomain.js
@@ -1,10 +1,5 @@
 'use strict';
 
-/* eslint-disable
-  no-nested-ternary,
-  multiline-ternary,
-  space-before-function-paren
-*/
 const url = require('url');
 const ip = require('internal-ip');
 
@@ -14,6 +9,7 @@ function createDomain(options, server) {
     ? ip.v4.sync() || 'localhost'
     : options.host;
 
+  // eslint-disable-next-line no-nested-ternary
   const port = options.socket ? 0 : server ? server.address().port : 0;
   // use explicitly defined public url
   // (prefix with protocol if not explicitly given)
