+++ lib/utils/createLogger.js
@@ -1,8 +1,5 @@
 'use strict';
 
-/* eslint-disable
-  space-before-function-paren
-*/
 const log = require('webpack-log');
 
 function createLogger(options) {
