+++ lib/utils/defaultTo.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function defaultTo(value, def) {
+  return value == null ? def : value;
+}
+
+module.exports = defaultTo;
