+++ lib/utils/status.js
@@ -1,44 +1,7 @@
 'use strict';
 
-/* eslint-disable
-  no-shadow,
-  global-require,
-  multiline-ternary,
-  array-bracket-spacing,
-  space-before-function-paren
-*/
 const open = require('opn');
-
-const colors = {
-  info(useColor, msg) {
-    if (useColor) {
-      // Make text blue and bold, so it *pops*
-      return `\u001b[1m\u001b[34m${msg}\u001b[39m\u001b[22m`;
-    }
-
-    return msg;
-  },
-  error(useColor, msg) {
-    if (useColor) {
-      // Make text red and bold, so it *pops*
-      return `\u001b[1m\u001b[31m${msg}\u001b[39m\u001b[22m`;
-    }
-
-    return msg;
-  },
-};
-
-// eslint-disable-next-line
-const defaultTo = (value, def) => {
-  return value == null ? def : value;
-};
-
-function version() {
-  return (
-    `webpack-dev-server ${require('../package.json').version}\n` +
-    `webpack ${require('webpack/package.json').version}`
-  );
-}
+const colors = require('./colors');
 
 function status(uri, options, log, useColor) {
   const contentBase = Array.isArray(options.contentBase)
@@ -96,27 +59,4 @@ function status(uri, options, log, useColor) {
   }
 }
 
-function bonjour(options) {
-  const bonjour = require('bonjour')();
-
-  bonjour.publish({
-    name: 'Webpack Dev Server',
-    port: options.port,
-    type: 'http',
-    subtypes: ['webpack'],
-  });
-
-  process.on('exit', () => {
-    bonjour.unpublishAll(() => {
-      bonjour.destroy();
-    });
-  });
-}
-
-module.exports = {
-  status,
-  colors,
-  version,
-  bonjour,
-  defaultTo,
-};
+module.exports = status;
